cast reduce_sum input to float64 to prevent overflow

+++ tensorboard/plugins/histogram/summary_v2.py
@@ -220,7 +220,11 @@ def _buckets(data, bucket_count=None):
             else non_singular_bucket_indices
         )
         clamped_indices = tf.minimum(bucket_indices, bucket_count - 1)
-        one_hots = tf.one_hot(clamped_indices, depth=bucket_count)
+        # Cast to float64 to prevent overflow beyond limt 2^24 in reduce_sum below.
+        one_hots = tf.cast(
+            tf.one_hot(clamped_indices, depth=bucket_count),
+            dtype=tf.float64,
+        )
         bucket_counts = tf.cast(
             tf.reduce_sum(input_tensor=one_hots, axis=0),
             dtype=tf.float64,
